% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_scores_to_tSNE_results.R
\name{add_scores_to_tSNE_results}
\alias{add_scores_to_tSNE_results}
\title{Add scores to list of t-SNE results.}
\usage{
add_scores_to_tSNE_results(
  tSNE_results_with_annotations,
  perplexityValues,
  scores,
  scoresName
)
}
\arguments{
\item{tSNE_results_with_annotations}{List of data frames. Each data frame column 1 and 2 should be
Y1 and Y2 t-SNE info respectively and additional columns containing t-SNE scores/annotations e.g.
stemness score or total transcription; row names should be single-cell IDs.}

\item{perplexityValues}{List of numbers corresponding to the perplexity value used for each t-SNE
in tSNE_results_with_annotations.}

\item{scores}{List of scores to add to tSNE_results_with_annotations.}

\item{scoresName}{Name to use for this score in tSNE_results_with_annotations.}
}
\value{
List of dataframes, i.e. tSNE_results_with_annotations but with score now added as an additional
field.
}
\description{
This function adds scores to a list of tSNE results for later printing e.g. add stemness predictions,
total transcription, total transcribed genes, etc.
}
