% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_zeros_to_perplexity_values.R
\name{add_zeros_to_perplexity_values}
\alias{add_zeros_to_perplexity_values}
\title{Function to add zeros to the start of perplexity values to make them nicer for printing.}
\usage{
add_zeros_to_perplexity_values(perplexityVals)
}
\arguments{
\item{Vector}{of numbers (perplexity values) to have zeros added to if they are <100.}
}
\value{
Vector of numbers with zeros now added.
}
\description{
This function takes a vector of numbers and adds zeros to the start of each to make
them nicer for printing.
}
