% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_matrix_conversions.R
\name{count_fractions_per_base_transcriptwise}
\alias{count_fractions_per_base_transcriptwise}
\title{Convert a matrix of read counts to a matrix of Cells Per Million values.}
\usage{
count_fractions_per_base_transcriptwise(counts, lengths)
}
\arguments{
\item{counts}{Expression matrix describing a cancer single-cell population; column
names should be gene names, row names should be cancer cell names.}

\item{lengths}{Vector containing the lengths of all the genes or their transcripts.}
}
\value{
Expression matrix containing Cells Per Million values.
}
\description{
This function converts an expression matrix (e.g. raw read counts), to
a matrix of Cells Per Million. The CellsPM matrix can be helpful in identifying
preferential expression e.g. in a group of cancer stem cells compared to a
tumour population with/without cells of others types.
}
