% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_matrix_conversions.R
\name{count_fractions_transcriptwise}
\alias{count_fractions_transcriptwise}
\title{Convert a matrix of expression values to a matrix of count fraction values.}
\usage{
count_fractions_transcriptwise(counts)
}
\arguments{
\item{counts}{Expression matrix describing a cancer single-cell population; column
names should be gene names, row names should be cancer cell names.}
}
\value{
Expression matrix containing count fraction values.
}
\description{
This function converts an expression matrix (e.g. of TPM values), to
a matrix of count fractions (e.g. TPM Count Fractions, if the input was TPMs). The
count fraction matrix can be helpful in identifying preferential expression e.g. in
a group of cancer stem cells compared to a tumour population with/without cells of
others types.
}
