% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_expression_scores_correlations.R
\name{get_expression_scores_correlations}
\alias{get_expression_scores_correlations}
\title{Get correlations between a single-cell expression matrix and a vector of scores for each cell.}
\usage{
get_expression_scores_correlations(tExprMatrix, scores)
}
\arguments{
\item{tExprMatrix}{Data frame describing t-SNE or PCA and score/s including the score to print.}

\item{scores}{Vector of scores to correlate with each row of tExprMatrix i.e. the expression
levels for all genes for each single-cell.}
}
\value{
List containing two named vectors that respectively contain the positive & negative
correlation results.
}
\description{
This function uses default correlation (Pearson) to give correlation values between a single-cell
RNA-Seq expression matrix and a vector of scores which contains one score for each cell.
For example, this could be a vector of predicted stemness scores, or a vector of
transcription summary values e.g. total transcribed genes, or total transcription for each cell.
}
