% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_or_transcript_length_fns.R
\name{get_gene_lengths_using_gene_names}
\alias{get_gene_lengths_using_gene_names}
\title{Get gene lengths using HUGO gene names.}
\usage{
get_gene_lengths_using_gene_names(gene_names, type = NULL)
}
\arguments{
\item{gene_names}{Vector of gene names to obtain the gene lengths for.}
}
\value{
Returned value genes_lengths is a named vector of integers containing
the full gene lengths for the continuous region between the chromsomal start and
stop locations provided by biomaRt/Ensembl.
}
\description{
This function uses biomaRt to obtain chromosomomal location which are used to obtain the
full gene lengths for each of a vector of gene names i.e. HGNC symbols.
}
