% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_HUGO_names_from_gene_IDs.R
\name{get_HUGO_names_from_gene_IDs}
\alias{get_HUGO_names_from_gene_IDs}
\title{Get HUGO gene names for a list of Ensembl gene IDs.}
\usage{
get_HUGO_names_from_gene_IDs(gene_IDs)
}
\arguments{
\item{gene_IDs}{List of gene IDs; an example of an Ensembl gene ID is 'ENSG00000000003'.}
}
\value{
A dataframe where column 1 = ensembl gene IDs and column 2 = hgnc_symbols.
}
\description{
This function uses biomaRt to look up HUGO gene names for a list of Ensembl gene IDs.
}
