% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_or_transcript_length_fns.R
\name{get_transcript_lengths_using_gene_names}
\alias{get_transcript_lengths_using_gene_names}
\title{Get transcript lengths using HUGO gene names.}
\usage{
get_transcript_lengths_using_gene_names(gene_names, type = NULL)
}
\arguments{
\item{gene_names}{Vector of gene names to obtain the transcript lengths for.}

\item{type}{Optional: 4 possible choices are:
1) "oldest" = get the lengths of the oldest transcript lengths - this uses the external
transcript names to indicate which transcript was first publicly deposited e.g. out of A1BG-201,
A1BG-202, A1BG-203, A1BG-204 and A1BG-205, the oldest transcript is taken as A1BG-201.
2) "longest" = get always the longest transcript out of all possibilites.
3) "appris or oldest" = if any of the transcripts for a gene has a BioMart APPRIS annotation
of 'principal1' use that transcript length, or if more than one has that annotation use the
average of them, otherwise use the oldest transcript. This is the default.
4) "appris or longest" = use transcripts annotated by APPRIS as 'principal1' otherwise
use the longest transcript.}
}
\value{
Returned value transcript_lengths is a named vector of integers.
}
\description{
This function uses biomaRt to obtain transcript lengths for each of a vector of gene names
(HGNC symbols).
}
