% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genes_full_names_fns.R
\name{getGenesAndFullNamesDF}
\alias{getGenesAndFullNamesDF}
\title{Convert a vector of gene names into a data frame containing gene names and matching full HUGO gene names.}
\usage{
getGenesAndFullNamesDF(
  genes_names,
  all_HUGO_abbreviations_and_fullnames = NULL,
  filesDir = NULL
)
}
\arguments{
\item{gene_names}{Vector of gene names to be converted into full gene names.}
}
\value{
Returned value genes_full_names is a data frame containing the gene names in the first column
and the full genes names in the second column.
}
\description{
This function uses hgnc_complete_set.txt from the EBI to convert a vector of gene names
into a DF containing the gene names and their corresponding full HUGO gene names (or '-' if not found).
}
