% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genes_full_names_fns.R
\name{getGenesFullNames}
\alias{getGenesFullNames}
\title{Convert a vector of gene names into their corresponding full HUGO gene names.}
\usage{
getGenesFullNames(
  genes_names,
  all_HUGO_abbreviations_and_fullnames = NULL,
  filesDir = NULL
)
}
\arguments{
\item{gene_names}{Vector of gene names to be converted into full gene names.}
}
\value{
Returned value genes_full_names is a vector of full genes names for the gene_names
parameter provided.
}
\description{
This function uses hgnc_complete_set.txt from the EBI to convert a vector of gene names
into their corresponding full HUGO gene names.
}
