% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_gene_positions_to_HUGO_gene_groups.R
\name{map_gene_positions_to_HUGO_gene_groups}
\alias{map_gene_positions_to_HUGO_gene_groups}
\title{For all HUGO gene groups find all their members positions in a list of genes.}
\usage{
map_gene_positions_to_HUGO_gene_groups(gene_names, HUGO_groups_list)
}
\arguments{
\item{gene_names}{List of HUGO gene names, e.g. list of column names in an scRNA-Seq
expression matrix.}

\item{HUGO_groups_list}{Named list of vectors, with each vector describing a HUGO gene group
by containing the HUGO gene symbols for each groups' members.}
}
\value{
HUGO_groups_positions Named list of vectors, with each vector containing the
positions of the HUGO gene group members in the gene_names list.
}
\description{
This function takes a list of gene names, and a named list of vectors, with each
vector describing a HUGO gene group by containing the HUGO gene symbols for each
groups' members. The function converts the named list of vectors, so that the gene names are
replaced with their corresponding positions in the gene list provided as the first param.
This mapping of HUGO gene groups to their positions in the list of genes is
used in the function printExpressionBarGraphs_for_families to produce plots of preferential
expression of HUGO gene groups.
}
