% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scores_and_fields.R
\name{plot_categories_with_grid}
\alias{plot_categories_with_grid}
\title{Plot a cell category overlaid on a supplied t-SNE or PCA generated using a cancer
single-cell population expression matrix, with grid in background.}
\usage{
plot_categories_with_grid(
  myData,
  xAxisLabel,
  yAxisLabel,
  runName,
  categoryName,
  datasetName,
  outDir,
  labelPositions = NULL,
  colourPositions = NULL,
  labelsVec = NULL,
  labelField = NULL
)
}
\arguments{
\item{myData}{Data frame describing t-SNE or PCA and additional info including the category to print.}

\item{runName}{Run name, e.g. 'tSNE_070'; will be used in the title of the plot and the output tif name.}

\item{categoryName}{Name of column in myData that contains the category info for plotting e.g. 'cell_type'.}

\item{datasetName}{Name of dataset e.g. 'mela_4558_TPM'.}

\item{outDir}{Output directory to use.}

\item{labelPositions}{Optional: Row number/s of myData i.e. cell positions, to print cell labels for.}

\item{colourPositions}{Optional: Row number/s of myData i.e. cell positions, to highlight cells (in black).}

\item{labelsVec}{Optional: Vector containing custom labels to display in legend for the printed category.}

\item{labelField}{Optional: If labelPositions parameter has been provided, labelField can provide the name of
another categorical variable within a column of myData to label the defined labelPositions.}

\item{xAxisName}{Name of column in myData that describes x-axis info for plot e.g. "Y1" in a t-SNE.}

\item{yAxisName}{Name of column in myData that describes y-axis info for plot e.g. "Y2" in a t-SNE.}
}
\value{
No return value.
}
\description{
This function is a generic function for plotting a t-SNE or PCA previously generated using
cancer single-cell population expression matrix. It can be used to label or colour
individual or groups of cells by categories of information (fields containing categorical variables/factors),
e.g. to assist identification of subpopulations of interest such as groups of cancer stem cells.
}
