% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scores_and_fields.R
\name{plot_field_with_thresholds}
\alias{plot_field_with_thresholds}
\title{Plot a score overlaid on a supplied t-SNE or PCA generated using a cancer single-cell population
expression matrix; a range in the score defined by 2 thresholds is coloured grey and light grey below
the lower threshold.}
\usage{
plot_field_with_thresholds(
  myData,
  xAxisName,
  yAxisName,
  fieldName,
  fullFieldName,
  thresholdOne,
  thresholdTwo,
  runName,
  outDir,
  labelPositions = NULL,
  colourPositions = NULL,
  labelField = NULL,
  highestOnTop = NULL,
  point_size = NULL
)
}
\arguments{
\item{myData}{Data frame describing t-SNE or PCA and score/s including the score to print.}

\item{xAxisName}{Name of column in myData that describes x-axis info for plot e.g. "Y1" in a t-SNE.}

\item{yAxisName}{Name of column in myData that describes y-axis info for plot e.g. "Y2" in a t-SNE.}

\item{fieldName}{Name of column in myData that contains the field for colouring the plot with.}

\item{fullFieldName}{Full name of fieldName to use in the plot title e.g. full HUGO gene name.}

\item{thresholdOne}{Lower value to print cells in grey. Cells <= this are in light grey.}

\item{thresholdTwo}{Upper value to print cells in grey. Cells > thresholdOne && < thresholdTwo are grey.}

\item{runName}{Run name, e.g. 'tSNE_070'; will be used in the title of the plot and the output tif name.}

\item{outDir}{Output directory to use.}

\item{labelPositions}{Optional: Row number/s of myData i.e. cell positions, to print cell labels for.}

\item{colourPositions}{Optional: Row number/s of myData i.e. cell positions, to highlight cells (in black).}

\item{labelField}{Optional: If labelPositions parameter has been provided, labelField can provide the name of
another categorical variable within a column of myData to label the defined labelPositions. E.g.
use labelPositions to define malignant cell clusters to be labelled, and use labelField to
label those cells with additional clinical info if available such as patient age, sex, or treament/s given.}

\item{highestOnTop}{Optional: Set to TRUE to print highest scoring cells on top of lower scoring cells.}

\item{point_size}{Optional: Size of points in plot.}
}
\value{
No return value.
}
\description{
For printing tSNEs or PCAs with additional field (e.g. individual gene) specified by
fieldName overlaid with an additional colour range in grey specified by thresholdOne &
thresholdTwo parameters (e.g. useful for lower gene expression range such as 0 to 1).
In grey >thresholdOne && < thresholdTwo; lgrey <= thresholdOne; coloured >= thresholdTwo.
It can be also be used to label or colour individual or small groups of cells to assist
identification of subpopulations of interest e.g. groups of cancer stem cells.
}
