% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scores_and_fields.R
\name{plot_scores_with_grid}
\alias{plot_scores_with_grid}
\title{Plot a score overlaid on a supplied t-SNE or PCA generated using a cancer
single-cell population expression matrix, with grid in background.}
\usage{
plot_scores_with_grid(
  myData,
  xAxisName,
  yAxisName,
  runName,
  scoresName,
  datasetName,
  outDir,
  labelPositions = NULL,
  colourPositions = NULL,
  highestOnTop = NULL
)
}
\arguments{
\item{myData}{Data frame describing t-SNE or PCA and additional info including the score to print.}

\item{xAxisName}{Name of column in myData that describes x-axis info for plot e.g. "Y1" in a t-SNE.}

\item{yAxisName}{Name of column in myData that describes y-axis info for plot e.g. "Y2" in a t-SNE.}

\item{runName}{Run name, e.g. 'tSNE_070'; will be used in the title of the plot and the output tif name.}

\item{scoresName}{Name of column in myData that contains the score info for plotting e.g. 'ranger_hipsci_raw_IPS'.}

\item{datasetName}{Name of dataset e.g. 'mela_4558_TPM'.}

\item{outDir}{Output directory to use.}

\item{labelPositions}{Optional: Row number/s of myData i.e. cell positions, to print cell labels for.}

\item{colourPositions}{Optional: Row number/s of myData i.e. cell positions, to highlight cells (in black).}

\item{highestOnTop}{Optional: Set to TRUE to print highest scoring cells on top of lower scoring cells.}
}
\value{
No return value.
}
\description{
This function is a generic function for plotting a t-SNE or PCA previously generated using
cancer single-cell population expression matrix. The grid in the background can help when
looking at exact positions of points (cells). It can also be used to label or colour
individual or small groups of cells to assist identification of subpopulations of interest
e.g. groups of cancer stem cells.
}
