% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_preferential_expression_bar_graphs.R
\name{print_preferential_expression_bar_graphs}
\alias{print_preferential_expression_bar_graphs}
\title{Print preferential expression bar graphs for subset of single-cells.}
\usage{
print_preferential_expression_bar_graphs(
  pCSC_group_name,
  pCSC_group_positions,
  HUGO_groups_sets_of_columnNumbers,
  tExprMatr,
  tumour_only_positions,
  this_patient_tumour_only_positions = NULL,
  outDir,
  tissue_type,
  family_type_name,
  patient_source_name,
  frag_count_method,
  no_ribo = NULL,
  separate_row_for_high_expression_transcripts = NULL,
  max_number_of_graph_rows_stacked_transcripts = NULL,
  max_number_of_graph_rows_stacked_families = NULL,
  HUGO_abbreviations_and_fullnames = NULL,
  HUGOFullNameFilesDir = NULL
)
}
\arguments{
\item{pCSC_group_name}{Name of CSC group to print preferential expression plots for
e.g. 'pCSC_group_1_SC4_Talazoparib.LB17011', but can be anything.}

\item{pCSC_group_positions}{Row number of CSCs in the CSC group in the expression matrix
tExprMatr.}

\item{HUGO_groups_sets_of_columnNumbers}{List of lists, i.e. a list of HUGO gene groups,
with each named gene group containing positions (row numbers) in the expression matrix
tExprMatr of all the gene group members present in the expression matrix.}

\item{tExprMatr}{Expression matrix describing a cancer single-cell population; column
names should be gene names, row names should be cancer cell names.}

\item{tumour_only_positions}{Positions (row numbers) in the expression matrix tExprMatr
thought to be tumour cells e.g. when the population contains cells annotated or predicted
as other cell types a subset may be defined, or if all cells are thought to be malignant
cancer cells this may be all cells.}

\item{this_patient_tumour_only_positions}{Optional: Expression matrix tExprMatr positions
(row numbers) for tumour cells from a specific patient i.e. that the CSCs for the
preferential expression plots are taken from. So, the plot needs to fulfil the criteria
that the CSCs are from only one patient, but the tumour cell population is from multiple
patients.}

\item{outDir}{Output directory for plots.}

\item{tissue_type}{Name of tissue type e.g. 'CDX'; can be anything.}

\item{family_type_name}{Name of the gene group type; usually 'HUGO'.}

\item{patient_source_name}{Optional: Name of the patient sample that CSC cells are from;
just name for in the plot title, can be anything.}

\item{frag_count_method}{Expression measurement type, e.g. 'raw'; but just a label, can be anything.}

\item{no_ribo}{Optional: Set to 1 to remove ribosomal transcripts from some of the bar charts}

\item{separate_row_for_high_expression_transcripts}{Optional: Set to 1 to print very high
expression transcripts or gene groups on a separate row in some of the bar charts; helpful
when there is a big difference between the most highly expressed and those on the same
bar chart row.}

\item{max_number_of_graph_rows_stacked_transcripts}{Optional: Set to manually override the
maximum number of rows shown in the stacked transcript plot.}

\item{max_number_of_graph_rows_stacked_families}{Optional: Set to manually override the
maximum number of rows shown in the stacked family plot.}
}
\value{
List of highest ranked preferentially expressed genes and gene families:
1 = ranked transcripts from single transcript barcharts;
2 = ranked families from family barcharts;
3 = ranked transcripts from family barcharts.
}
\description{
This function prints multiple bar graphs to display preferential expression
in a defined subset of cells e.g. a group of cancer stem cells.
}
