% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scores_and_fields.R
\name{print_tSNE_coloured_by_gene}
\alias{print_tSNE_coloured_by_gene}
\title{Print a list of t-SNEs each overlaid with expression of a specic genes; multiple t-SNE prints can generated
using a list of genes.}
\usage{
print_tSNE_coloured_by_gene(
  gene_names_to_print,
  tSNE_with_annotations,
  tExprMatr,
  run_name,
  outDir,
  point_size = NULL,
  HUGO_abbreviations_and_fullnames = NULL,
  HUGOFullNameFilesDir = NULL,
  use_biomart = FALSE
)
}
\arguments{
\item{gene_names_to_print}{List of one or more gene names to print expression levels for on one or more t-SNEs.}

\item{tSNE_with_annotations}{Each data frame column 1 and 2 should be Y1 and Y2 t-SNE info respectively and
additional columns containing t-SNE scores/annotations e.g. stemness score or total transcription; row names
should be single-cell IDs.}

\item{tExprMatr}{List of expression matrices describing cancer single-cell population; for each
matrix, column names should be gene names, and row names should be cell IDs.}

\item{run_name}{Run name, e.g. 'tSNE_070'; will be used in the title of the plot and the output tif name.}

\item{outDir}{Output directory to use.}

\item{point_size}{Size of points in plot.
value used.}
}
\value{
No return value.
}
\description{
This function is a wrapper for calling plotFieldWithThresholds to print a list of t-SNEs described
in tSNE_list_with_annotations, each overlaid with expression for one gene taken from CDX_tRawCounts_list.
}
