% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_preferential_expression_bar_graphs.R
\name{printBarchartSingleTranscriptNoStacking}
\alias{printBarchartSingleTranscriptNoStacking}
\title{Print expression bar graph for genes/transcripts most highly expressed in selected
putative CSCs.}
\usage{
printBarchartSingleTranscriptNoStacking(
  tExprMatr,
  pCSC_group_positions,
  pCSC_group_name,
  outDir,
  tissue_type,
  family_type_name,
  patient_source_name,
  no_ribo = TRUE,
  number_of_transcripts_to_show = NULL,
  HUGO_abbreviations_and_fullnames = NULL,
  HUGOFullNameFilesDir = NULL
)
}
\arguments{
\item{tExprMatr}{Expression matrix describing a cancer single-cell population; column
names should be gene names, row names should be cancer cell names.}

\item{pCSC_group_positions}{Row number of CSCs in the CSC group in the expression matrix
tExprMatr.}

\item{pCSC_group_name}{Name of CSC group to print preferential expression plots for
e.g. 'pCSC_group_1_SC4_Talazoparib.LB17011', but can be anything.}

\item{outDir}{Output directory for plots.}

\item{tissue_type}{Name of tissue type e.g. 'CDX'; can be anything.}

\item{family_type_name}{Name of the gene group type; usually 'HUGO'.}

\item{patient_source_name}{Optional: Name of the patient sample that CSC cells are from;
just name for in the plot title, can be anything.}

\item{no_ribo}{Optional: Set to 1 to remove ribosomal transcripts from some of the bar charts}

\item{number_of_transcripts_to_show}{Optional: Number of transcripts to include.}
}
\value{
Nothing returned.
}
\description{
This function prints a bar graph to display expression in the genes most highly expressed
in a selected set of putative CSCs. The order is by mean transcript expression in the
pCSC cells i.e. highest first. If the no_ribo parameter is used, the same bar
graph is printed with all transcripts with "ribosomal" in their full name removed
from the results.
}
