% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_preferential_expression_bar_graphs.R
\name{printTripletSingleTranscriptBarcharts}
\alias{printTripletSingleTranscriptBarcharts}
\title{Print expression bar graph for genes/transcripts most highly expressed in selected
putative CSCs vs tumour cells vs all cells.}
\usage{
printTripletSingleTranscriptBarcharts(
  tExprMatr,
  pCSC_group_positions,
  tumour_only_positions,
  pCSC_group_name,
  outDir,
  tissue_type,
  family_type_name,
  patient_source_name,
  no_ribo = TRUE,
  number_of_transcripts_to_show = NULL,
  HUGO_abbreviations_and_fullnames = NULL,
  HUGOFullNameFilesDir = NULL
)
}
\arguments{
\item{tExprMatr}{Expression matrix describing a cancer single-cell population; column
names should be gene names, row names should be cancer cell names.}

\item{pCSC_group_positions}{Row number of CSCs in the CSC group in the expression matrix
tExprMatr.}

\item{tumour_only_positions}{Positions (row numbers) in the expression matrix tExprMatr
thought to be tumour cells e.g. when the population contains cells annotated or predicted
as other cell types a subset may be defined, or if all cells are thought to be malignant
cancer cells this may be all cells.}

\item{pCSC_group_name}{Name of CSC group to print preferential expression plots for
e.g. 'pCSC_group_1_SC4_Talazoparib.LB17011', but can be anything.}

\item{outDir}{Output directory for plots.}

\item{tissue_type}{Name of tissue type e.g. 'CDX'; can be anything.}

\item{family_type_name}{Name of the gene group type; usually 'HUGO'.}

\item{patient_source_name}{Optional: Name of the patient sample that CSC cells are from;
just name for in the plot title, can be anything.}

\item{no_ribo}{Optional: Set to 1 to remove ribosomal transcripts from some of the bar charts}

\item{number_of_transcripts_to_show}{Optional: Number of transcripts to include.}
}
\value{
Nothing returned.
}
\description{
This function prints a bar graph to display relative expression in the genes most
highly expressed in a selected set of putative CSCs, along with their expression
in tumour cells and all cells. The order is by mean transcript expression in the
pCSC cells i.e. highest first. If the no_ribo parameter is used, the same bar
graph is printed with all transcripts with "ribosomal" in their full name removed
from the results.
}
