% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scores_and_fields.R
\name{round_down}
\alias{round_down}
\title{Function to round down numbers to nearest numbers divisible by a specified value.}
\usage{
round_down(value, by)
}
\arguments{
\item{value}{Value to round down}

\item{by}{Round down to the nearest value divisible by this number}
}
\value{
Rounded down value
}
\description{
This function rounds down to a nearest number, e.g. value = -7, by = 10, will round down to -10.
}
