% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scores_and_fields.R
\name{round_up}
\alias{round_up}
\title{Function to round up numbers to nearest numbers divisible by a specified value.}
\usage{
round_up(value, by)
}
\arguments{
\item{value}{Value to round up}

\item{by}{Round up to the nearest value divisible by this number}
}
\value{
Rounded up value
}
\description{
This function rounds up to a nearest number, e.g. value = 7, by = 10, will round up to 10.
}
