% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tSNEs.R
\name{run_all_tSNEs}
\alias{run_all_tSNEs}
\title{Run t-SNEs for cancer single-cells expression matrix for all perplexity values provided.}
\usage{
run_all_tSNEs(
  perplexityValues,
  tExpressionMatrix,
  runName,
  outDir,
  montage = FALSE
)
}
\arguments{
\item{perplexityValues}{List of perplexity values to use e.g. c(5, 10, 20, 30, 40, 50, 60, 70);
recommended = 70.}

\item{tExpressionMatrix}{Expression matrix for cancer single-cell population or sample; where
column names should be gene names, and row names should be sample names.}

\item{runName}{Label for run, e.g. 'rawCounts_GSM4104152_SC4.LB17009'; can be anything.}

\item{outDir}{Directory name for printing to.}

\item{montage}{If TRUE use montage system command to combine previously generated plots in tif
files.}
}
\value{
A list of data frames, each containing a t-SNE result.
}
\description{
This function runs one or more t-SNEs for a cancer single-cell expression matrix i.e. each
using a different perplexity value, provided perplexityValues; recommended = 70. This
function calls runTSNE which generates the t-SNE.
}
