% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ML_predictors.R
\name{run_ML_cellType_predictors}
\alias{run_ML_cellType_predictors}
\title{Run machine learning stemness predictors.}
\usage{
run_ML_cellType_predictors(
  model_type,
  sample_tCounts_to_use_list,
  cellTypeData,
  expressionData_type,
  number_of_genes_to_compare,
  sample_data_type,
  cellType_out_dir
)
}
\arguments{
\item{model_type}{Type of classifier to use: 'ranger', 'rf', 'randomForest' or 'gbm'.}

\item{sample_tCounts_to_use_list}{List of single-cell cancer samples in expression matrix where column headers are gene names, and
row names are cancer cell IDs.}

\item{cellTypeData}{Expression data for training where column one must be 'Cell_type' and the rest are gene names, and row names are sample names.}

\item{expressionData_type}{Name of expression data e.g. 'NSCLC', 'melanoma', 'hipsci', etc.; can be anything.}

\item{number_of_genes_to_compare}{Number of genes to compare between the samples and the expression data; recommended = 16000 to 10000.}

\item{sample_data_type}{Type of single-cell expression data being used e.g. 'raw' or 'tpm'; just a name though, can be anything.}

\item{cellType_out_dir}{Directory to save results from this function.}
}
\value{
A list of prediction results i.e. one for each scRNA-Seq sample.
}
\description{
This function runs ML classifiers on single-cell expression matrices to predict
biological characteristics using annotationed expression data as the training data
for the classifier. In particular, use this function to predict stemness using one
of three random forest methods ('ranger', 'rf', or 'randomForest') or gbm, training
them with expressionData with cell type annotations to predict cell type probability
for each cell in the cancer single-cell populations.
}
