% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ML_predictors.R
\name{run_ML_stemness_predictors}
\alias{run_ML_stemness_predictors}
\title{Run machine learning stemness predictors.}
\usage{
run_ML_stemness_predictors(
  model_type,
  sample_tCounts_to_use_list,
  stemData,
  stemness_type,
  number_of_genes_to_compare,
  sample_data_type,
  stemness_out_dir
)
}
\arguments{
\item{model_type}{Type of classifier to use: 'ranger', 'rf', or 'randomForest'.}

\item{sample_tCounts_to_use_list}{List of single-cell cancer samples in expression matrix where column headers are gene names, and
row names are cancer cell IDs.}

\item{stemData}{Stemness expression data for training where column one must be 'Cell_type' and the rest are gene names, and row names are sample names.}

\item{stemness_type}{Name of stemness data; choices are 'hipsci' or 'PCBC'.}

\item{number_of_genes_to_compare}{Number of genes to compare between the samples and the stemness data; recommended = 10000 to 16000.}

\item{sample_data_type}{Type of single-cell expression data being used e.g. 'raw' or 'tpm'; just a name though, can be anything.}

\item{stemness_out_dir}{Directory to save results from this function.}
}
\value{
A list of prediction results i.e. one for each scRNA-Seq sample.
}
\description{
This function runs ML classifiers on single-cell expression matrices to predict
biological characteristics using annotationed expression data as the training data
for the classifier. In particular, use this function to predict stemness using one
of three random forest methods ('ranger', 'rf', or 'randomForest'), training
them with either 'HipSci' or 'PCBC' stemness expression data to predict stemness
probability for each cell in the cancer single-cell populations.
}
