% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tSNEs.R
\name{run_tSNE}
\alias{run_tSNE}
\title{Run a t-SNE for cancer single-cells expression matrix for a perplexity value provided.}
\usage{
run_tSNE(perplexityValue, tExprMatr, runName, outDir)
}
\arguments{
\item{perplexityValue}{Perplexity value to use e.g. 5-130; recommended = 70.}

\item{tExprMatr}{Expression matrix for cancer single-cell population; where
column names should be gene names, and row names should be sample names.}

\item{runName}{Label for run, e.g. 'rawCounts_GSM4104152_SC4.LB17009'; can be anything.}

\item{outDir}{Directory name for printing to.}
}
\value{
A data frame describing the t-SNE result i.e. Y1 & Y2.
}
\description{
This function runs a t-SNE for a cancer single-cell expression matrix using a provided perplexity
value e.g. 70.
}
