% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_tSNE_box_positions.R
\name{select_tSNE_box_positions}
\alias{select_tSNE_box_positions}
\title{Select cells from a t-SNE that are located within a defined rectangle.}
\usage{
select_tSNE_box_positions(tSNE_info, Y1_borders, Y2_borders)
}
\arguments{
\item{tSNE_info}{Data frame where first 2 columns are Y1 and Y2 results from a
t-SNE generated using a cancer poulation expression matrix. Rows are cancer
cells. Additional columns may contain additional info, such as scores of
the cells or clinical info about the cells.}

\item{Y1_borders}{Opposite sides of the rectangle to locate cells within it, in
Y1 positions in a vector, e.g. Y1_borders = c(3,5).}

\item{Y2_borders}{Opposite sides of the rectangle in Y2 positions in a
vector, e.g. Y2_borders = c(2,4).}
}
\value{
Positions of the cells in the DF rows, if any, that are located in the
rectangle defined by Y1_borders and Y2_borders.
}
\description{
Identify and return the positions of cells that are located in the t-SNE within
the rectangle defined by parameters Y1_borders and Y2_borders. Can be used,
for example, to find the cells that make up a cluster of cells in the t-SNE.
}
