% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_preferential_expression_bar_graphs.R
\name{stackedBarchartsForSingleTranscripts}
\alias{stackedBarchartsForSingleTranscripts}
\title{Print expression stacked bar graph for genes/transcripts most preferentially expressed in selected
putative CSCs.}
\usage{
stackedBarchartsForSingleTranscripts(
  tExprMatr,
  pCSC_group_positions,
  tumour_only_positions,
  this_patient_tumour_only_positions = NULL,
  pCSC_group_name,
  tissue_type,
  family_type_name,
  patient_source_name,
  separate_row_for_high_expression_transcripts = NULL,
  no_ribo = NULL,
  manually_set_max_number_of_graph_rows = NULL,
  HUGO_abbreviations_and_fullnames = NULL,
  HUGOFullNameFilesDir = NULL
)
}
\arguments{
\item{tExprMatr}{Expression matrix describing a cancer single-cell population; column
names should be gene names, row names should be cancer cell names.}

\item{pCSC_group_positions}{Row number of CSCs in the CSC group in the expression matrix
tExprMatr.}

\item{tumour_only_positions}{Positions (row numbers) in the expression matrix tExprMatr
thought to be tumour cells e.g. when the population contains cells annotated or predicted
as other cell types a subset may be defined, or if all cells are thought to be malignant
cancer cells this may be all cells.}

\item{this_patient_tumour_only_positions}{Optional: Expression matrix tExprMatr positions
(row numbers) for tumour cells from a specific patient i.e. that the CSCs for the
preferential expression plots are taken from. So, the plot needs to fulfil the criteria
that the CSCs are from only one patient, but the tumour cell population is from multiple
patients.}

\item{pCSC_group_name}{Name of CSC group to print preferential expression plots for
e.g. 'pCSC_group_1_SC4_Talazoparib.LB17011', but can be anything.}

\item{tissue_type}{Name of tissue type e.g. 'CDX'; can be anything.}

\item{family_type_name}{Name of the gene group type; usually 'HUGO'.}

\item{patient_source_name}{Optional: Name of the patient sample that CSC cells are from;
just name for in the plot title, can be anything.}

\item{separate_row_for_high_expression_transcripts}{Optional: Set to 1 to print very high
expression transcripts or gene groups on a separate row in some of the bar charts; helpful
when there is a big difference between the most highly expressed and those on the same
bar chart row.}

\item{no_ribo}{Optional: Set to 1 to remove ribosomal transcripts from some of the bar charts}

\item{manually_set_max_number_of_graph_rows}{Optional: Set to manually override the
maximum number of rows shown in the plot.}
}
\value{
Highest ranked preferentially expressed genes
}
\description{
This function prints a stacked bar graph to display expression in the genes/transcripts most
preferentially expressed in a selected set of putative CSCs. This may use expression in the
form of read counts, TPM, TPM count fractions, Cells Per Million, etc. Cells Per Million or
TPM count fractions are recommended for tumour cell populations containing a mix of cell types.
Read counts are recommended for homogenous cell populations e.g. PDXs, CDXs. Each bar in the
plot generated is subdivided by the proportion of the average expression that is contributed
by each cell in the pCSC group. The bars are ordered by these average expression scores across
the pCSC group.
}
