% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_matrix_conversions.R
\name{tpm_samplewise}
\alias{tpm_samplewise}
\title{Convert a matrix of read counts to a matrix of Transcripts Per Million values.}
\usage{
tpm_samplewise(counts, lengths)
}
\arguments{
\item{counts}{Expression matrix e.g. describing a cancer single-cell population where
column names should be gene names, row names should be cancer cell names.}

\item{lengths}{Vector containing the lengths of all the transcripts.}
}
\value{
Expression matrix containing TPM values.
}
\description{
This function converts an expression matrix of raw read counts, to
a matrix of TPM values.
}
